Advance styling structure with the use of styled-components
Installed some more modules for google oath login and some styling ones to aid our understanding of the app's execution and help make designing easier. Furthermore advanced the app's design to approach our high fidelity-design. Users at this point are able to log in but no cookies are stored and nothing more can be done except logging out right after. The method of log-in might also change in the future, depending on functionality.