Update handleRadioChange and handleCheckboxChange
Both functions and their respective HTML are updated to be reused with any other possible HTML radio or checkbox input.

Explanation:
For radios, all radios need to provide the "other" option in their onChange in order to reset it whenever one of them is clicked.
For checkboxes, only the other option should delete it so it is the only one where the custom handler will be used. All the rest can use formik.handleChange directly.