Major CSS improvement and form steps with validation fix
Now one can navigate through the different pages of the form correctly since the handleNext function in Form.jsx now implements all other steps and checks their respective errors before advancing.

CSS is improved majorly around the entirety of the app to be easier on the eyes and achieve an aesthetic look. Also redundant code in Form.styled.js is removed. Ignore Pg0, it is to be fixed later on.