Hide Application contents, except for initial page if the user is not logged in. Same with Dashboard and My Applications
Dashboard and My Applications are no longer accessible URLs and do not appear in the sidebar if the user is not logged in. Additionally, the Application shows only the initial page which is just a guide and prompts the user to sign in when he tries to begin an application while not logged in.