Fix comments and applications fetch
Fix comments not being added 90% of the time because the window refreshed too quickly, and fix the applications fetch in dashboard so they all load quickly and efficiently. Once when the page loads and once again when buttons that don't lead to another page are pressed. Also redirect from application to dashboard if you just commented, and to my applications if you just submitted an application.