Application storyline established, dashboard upgraded, admin powers enhanced
The Dashboard now fetches applications where the logged in user is a reviewer in too, in addition to the supervisor ones which were already implemented.

The application now has stages and is unable to proceed unless a certain step has been taken. The current implementation takes into account 1 reviewer only for the low risk applications. Future commits will include the functionality of approval by more than one reviewers.

The admin functionality now also adds a user_role row when someone becomes an admin, and deletes it when they are changed to staff or student. This implementation will make it easier to distinguish who is accepting which application in each step and as which role, since we discussed a person with more permission might also have more roles in one application so will have to keep the application flow as it is, instead of skip ahead.