Change mode view to review, enhance navigation between reviewing and dashboard and insert missing styling
The mode named view is renamed to review since it is used to view and add comments to an application which someone can only access through the dashboard. The mode named view will be kept for the user since he can view the application. Allowed the navigation to go backwards from the application being reviewed, to the dashboard. Added some missing styling from the comment component.