Fix application story for admins
The admin can now be both a supervisor and a reviewer in the same application and attempt to approve an application, the storyline will be kept with proper prompts if shortcuts are attempted.

Also fixed the approval by admins irrelevant to an application being able to approve when they shouldn't.