Beginning of comments to the story
Now if comments exist, the edit application button turns on for the user so he can do the necessary changes. The comments can be seen inside the application under the form, initially. Next we should take into account which stage the application was in when the applicant makes the changes required so the path can continue as was.