Complete files upload, sorting and association with applications
Files are now uploaded and saved in separate folders for each application. The files are associated with each application via the folder name "application_id_X" and the files are associated with each field through the use of extra formik values which are created dynamically, which contain the file name and are inserted in the database in their respective fields. Next  step is to display the files when viewing an application, either as a link to a new tab, or the files themselves.